requirements.txt
streamlit
Pillow
opencv-python
exifread
numpy
matplotlib
